Refactor standalone code from osmosis code.
