+++ src/main/java/net/mojodna/osm2orc/osmosis/OrcWriterFactory.java
@@ -1,4 +1,4 @@
-package net.mojodna.osm2orc;
+package net.mojodna.osm2orc.osmosis;
 
 import org.openstreetmap.osmosis.core.pipeline.common.TaskConfiguration;
 import org.openstreetmap.osmosis.core.pipeline.common.TaskManager;
