+++ src/main/java/net/mojodna/osm2orc/standalone/OsmChangesetXml2Orc.java
@@ -1,0 +1,8 @@
+package net.mojodna.osm2orc.standalone;
+
+
+public class OsmChangesetXml2Orc {
+    public static void convert(String inputChangesetXml, String outputOrc) {
+
+    }
+}
