+++ src/main/resources/osmosis-plugins.conf
@@ -1,1 +1,1 @@
-net.mojodna.osm2orc.OrcPluginLoader
+net.mojodna.osm2orc.osmosis.OrcPluginLoader
